requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
djangorestframework==3.15.2
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
pytest==8.3.4
pytest-django==4.9.0
python-decouple==3.8
python3-openid==3.2.0
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
